Stage 1
Completed the project to an extent where it is able to connect to the board, specify if it is a cyton or a virtual simulation, save the data straight to a csv file, and plot the down sampled signal.

stage two will aim to:
- plot multiple channels as subplots
- paramterize some of the parts of the code
- make the plotted window time span a fixed length that advances over time

perhaps stage 3 will delve into real-time signal analysis, aiming to be able to extract features in real time froms treaming signals.